+++ docs/examples/brep_explorer.py
@@ -1,11 +1,10 @@
 from compas.geometry import Box
 from compas.colors import Color
-from compas_occ.brep import BRep
+from compas.geometry import Brep
 from compas_view2.app import App
-from compas_view2.objects import Collection
 
 box = Box.from_width_height_depth(1, 1, 1)
-A = BRep.from_box(box)
+A = Brep.from_box(box)
 A.sew()
 A.fix()
 
@@ -22,8 +21,7 @@ print(len(faces))
 
 viewer = App()
 
-viewmesh = A.to_viewmesh()
-viewer.add(Collection(viewmesh[1]), linewidth=1, linecolor=Color(0.2, 0.2, 0.2))
+viewer.add(A, linewidth=1, linecolor=Color(0.2, 0.2, 0.2))
 
 viewer.add(vertex.point, pointcolor=Color.red(), pointsize=20)
 
@@ -34,8 +32,8 @@ for edge in edges:
     viewer.add(edge.to_line(), linewidth=5, linecolor=Color(0.2, 0.2, 0.2))
 
 for face in faces:
-    brep = BRep()
-    brep.occ_shape = face.occ_face
-    viewer.add(brep.to_viewmesh()[0], show_lines=False, opacity=0.5)
+    brep = Brep()
+    brep.native_brep = face.occ_face
+    viewer.add(brep, show_lines=False, opacity=0.5)
 
 viewer.show()