+++ docs/examples/brep_from_booleans.py
@@ -1,6 +1,6 @@
 from compas.geometry import Point, Vector, Plane
 from compas.geometry import Box, Cylinder
-from compas_occ.brep import BRep
+from compas.geometry import Brep
 from compas_view2.app import App
 from compas_view2.objects import Collection
 
@@ -18,10 +18,10 @@ cx = Cylinder((YZ, 0.7 * R), 4 * R)
 cy = Cylinder((ZX, 0.7 * R), 4 * R)
 cz = Cylinder((XY, 0.7 * R), 4 * R)
 
-A = BRep.from_box(box)
-B1 = BRep.from_cylinder(cx)
-B2 = BRep.from_cylinder(cy)
-B3 = BRep.from_cylinder(cz)
+A = Brep.from_box(box)
+B1 = Brep.from_cylinder(cx)
+B2 = Brep.from_cylinder(cy)
+B3 = Brep.from_cylinder(cz)
 
 C = A - (B1 + B2 + B3)
 
@@ -29,19 +29,11 @@ C = A - (B1 + B2 + B3)
 # Visualisation
 # ==============================================================================
 
-# C.data = C.data
-
-# Currently, the viewer does not suppport BRep shapes.
-# Therefore we have to convert the components of the BRep to something the viewer does understand.
-
 viewer = App(viewmode="ghosted", width=1600, height=900)
 viewer.view.camera.rz = -30
 viewer.view.camera.rx = -75
 viewer.view.camera.distance = 7
 
-viewmesh = C.to_viewmesh()
-
-viewer.add(viewmesh[0], show_lines=False)
-viewer.add(Collection(viewmesh[1]), linewidth=2)
+viewer.add(C)
 
 viewer.run()