+++ src/compas_occ/conversions/primitives.py
@@ -12,6 +12,7 @@ from compas.geometry import Sphere
 from compas.geometry import Cylinder
 from compas.geometry import Cone
 from compas.geometry import Torus
+from compas.geometry import Transformation
 
 from OCC.Core.gp import gp_Ax1
 from OCC.Core.gp import gp_Ax2
@@ -29,6 +30,7 @@ from OCC.Core.gp import gp_Sphere
 from OCC.Core.gp import gp_Cylinder
 from OCC.Core.gp import gp_Cone
 from OCC.Core.gp import gp_Torus
+from OCC.Core.TopLoc import TopLoc_Location
 
 
 def compas_point_to_occ_point(point: Point) -> gp_Pnt:
@@ -364,6 +366,28 @@ def compas_plane_to_occ_ax3(plane: Plane) -> gp_Ax3:
         compas_point_to_occ_point(plane.point),
         compas_vector_to_occ_direction(plane.normal),
     )
+
+
+def compas_frame_from_location(location: TopLoc_Location):
+    """Convert an OCC Location to COMPAS Frame.
+
+    Parameters
+    ----------
+    location : ``TopLoc_Location``
+        The OCC Location to convert to frame.
+
+    Returns
+    -------
+    :class:`~compas.geometry.Frame`
+
+    """
+    t = location.Transformation()
+
+    # transformation.Value is a 1-based 3x4 matrix
+    rows, columns = 3, 4
+    matrix = [[t.Value(i, j) for j in range(1, columns + 1)] for i in range(1, rows + 1)]
+    matrix.append([0.0, 0.0, 0.0, 1.0])  # COMPAS wants a 4x4 matrix
+    return Frame.from_transformation(Transformation(matrix))
 
 
 def compas_frame_from_occ_ax3(position: gp_Ax3, cls: Type[Frame] = None) -> Frame: