+++ CHANGELOG.md
@@ -9,7 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `BRep.trim` and `BRep.trimmed`.
+* Added `BRep.sliced` as an operation on a copy of `BRep.slice`.
 * Added `compas_occ.geometry.OCCCurve.parameter_at_distance`.
+* Added `compas_occ.conversions.compas_frame_from_location`.
+
+### Changed
+
+* Changed `compas_occ.brep.BRep` to be a pluggin for `compas.geometry.Brep`.
+* Renamed `BRep.occ_shape` to `native_brep` for compatibility with `RhinoBrep`.
+* Fixed the error when calling `BRep.frame`.
+* Fixed `AttributeError` when calling `BRep.loops`.
 
 ### Changed